Update Data Loaders.md
Hello, pardon my intrusion.  
I thought to check this repo that you just made public and then found some styles that may not have been what you intended, and since I always wanted to contribute to a public repo that was not mine, I thought to give a hand.   
If this seems like help then you may commit it or if not you may disregard it.  
Happy holiday. 😄